# Conflicts:
#	openaev-api/src/main/java/io/openaev/api/chaining/InjectExecutionStep.java
#	openaev-api/src/main/java/io/openaev/service/InjectExpectationService.java
#	openaev-api/src/main/java/io/openaev/service/chaining/WorkflowService.java
#	openaev-api/src/test/java/io/openaev/service/chaining/StepServiceIntegrationTest.java

+++ openaev-api/src/test/java/io/openaev/service/ExerciseServiceIntegrationTest.java
@@ -9,6 +9,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 import io.openaev.IntegrationTest;
+import io.openaev.api.url_access_token.UrlAccessTokenService;
 import io.openaev.config.cache.LicenseCacheManager;
 import io.openaev.database.model.*;
 import io.openaev.database.repository.*;
@@ -84,6 +85,7 @@ class ExerciseServiceIntegrationTest extends IntegrationTest {
   @Autowired private LessonsService lessonsService;
   @Autowired private FileService fileService;
   @Autowired private PauseExerciseService pauseExerciseService;
+  @Autowired private UrlAccessTokenService urlAccessTokenService;
 
   @Autowired private WorkflowService workflowService;
   @Autowired private io.openaev.healthcheck.utils.HealthCheckUtils healthCheckUtils;
@@ -129,6 +131,7 @@ class ExerciseServiceIntegrationTest extends IntegrationTest {
             lessonsAnswerRepository,
             lessonsCategoryRepository,
             lessonsService,
+            urlAccessTokenService,
             injectExpectationMapper,
             scenarioRecurrenceService,
             workflowService,
@@ -141,9 +144,9 @@ class ExerciseServiceIntegrationTest extends IntegrationTest {
 
   @AfterAll
   public void teardown() {
-    this.userRepository.deleteById(USER_ID);
-    this.teamRepository.deleteById(TEAM_ID);
-    this.injectRepository.deleteById(INJECT_ID);
+    if (USER_ID != null) this.userRepository.deleteById(USER_ID);
+    if (TEAM_ID != null) this.teamRepository.deleteById(TEAM_ID);
+    if (INJECT_ID != null) this.injectRepository.deleteById(INJECT_ID);
   }
 
   @DisplayName("Should create new contextual teams while exercise duplication")
@@ -156,7 +159,10 @@ class ExerciseServiceIntegrationTest extends IntegrationTest {
     exerciseTeams.add(contextualTeam);
     Team noContextualTeam = this.teamRepository.save(getTeam(null, "fakeTeamName2", false));
     exerciseTeams.add(noContextualTeam);
-    Exercise exercise = this.exerciseRepository.save(getExercise(exerciseTeams));
+    Exercise exercise = getExercise(exerciseTeams);
+    exercise.setFrom("test@test.com");
+    this.exerciseRepository.save(exercise);
+    entityManager.flush();
 
     // -- EXECUTE --
     Exercise exerciseDuplicated = exerciseService.getDuplicateExercise(exercise.getId());
